fix: memleaks in child fork proc
