+++ Makefile
@@ -15,7 +15,7 @@ PARSER_FILES	=	syntax_parser parser_data parse_redirection build_parser_tree
 LEXER_FILES	=	lexer lexer_data tokenizer
 
 HELPERS_FILES = helpers modify_cmd modify_cmd_helpers modify_cmd_helpers2 \
-	mutate_cmd destroy_data
+	mutate_cmd destroy_data get_path
 
 EXEC_FILES	= executor execute_process redirect redir_utils heredoc_redir \
 	execute_process_utils
