+++ includes/mini_shell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   mini_shell.h                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tklimova <tklimova@student.42.fr>          +#+  +:+       +#+        */
+/*   By: tklimova <tklimova@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 14:58:51 by tklimova          #+#    #+#             */
-/*   Updated: 2023/12/12 14:31:22 by tklimova         ###   ########.fr       */
+/*   Updated: 2023/12/12 14:05:13 by tklimova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -110,6 +110,7 @@ typedef struct s_exec_data
 	int		go_on;
 	int		fd_out;
 	char	*err_file;
+	t_data	*link_to_data;
 }			t_exec_data;
 
 typedef struct s_gb
