+++ src/executor/execute_process_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   execute_process_utils.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tklimova <tklimova@student.42.fr>          +#+  +:+       +#+        */
+/*   By: tklimova <tklimova@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/12 14:15:20 by tklimova          #+#    #+#             */
-/*   Updated: 2023/12/12 15:08:37 by tklimova         ###   ########.fr       */
+/*   Updated: 2023/12/12 12:43:46 by tklimova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,4 +35,5 @@ void	bind_current_path_to_cmd(t_parser_data *parser_node, t_env *env)
 	free(parser_node->cmd_line[0]);
 	parser_node->cmd_line[0] = ft_strcopy(cmd_with_path);
 	free(cmd_with_path);
+	printf("\n ARFER MUTATION parser_node->text: %s\n", parser_node->text);
 }
