+++ src/lexer/lexer.c
@@ -6,7 +6,7 @@
 /*   By: tklimova <tklimova@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/15 13:28:25 by root              #+#    #+#             */
-/*   Updated: 2023/12/11 17:46:52 by tklimova         ###   ########.fr       */
+/*   Updated: 2023/12/12 13:54:44 by tklimova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -92,5 +92,6 @@ void	lexer(t_data *data, char *cmd_buff)
 		cmd_buff += w_len;
 	}
 	syntax_parser(data);
+	printf("\nDESTROY!!!!\n");
 	destroy_data(data);
 }
